backend/requirements.txt
Django
djangorestframework
psycopg2-binary
python-decouple
drf-yasg
django-cors-headers
drf-spectacular
channels
channels-redis
daphne